mcp/document-db/pyproject.toml
[project]
name = "document-db-mcp"
version = "0.1.0"
description = "Document DB MCP — PostgreSQL JSONB CRUD over streamable HTTP for dev-team agents"
requires-python = ">=3.13,<3.14"

# 모든 하한은 PyPI 최신 stable (2026-05 시점) 기준
dependencies = [
  # MCP SDK — streamable HTTP transport (spec 2025-06-18)
  "mcp>=1.27.0",
  # PostgreSQL async driver (ORM 도입 금지 — root CLAUDE.md / mcp/document-db/CLAUDE.md §8)
  "asyncpg>=0.31.0",
  # Schema validation
  "pydantic>=2.13.3",
  # env / config
  "pydantic-settings>=2.14.0",
  # Migration (CLAUDE.md §6)
  "yoyo-migrations>=9.0.0",
  # yoyo 의 sync driver — psycopg3 (langgraph 와 동일 driver)
  "psycopg[binary]>=3.2.0",
  # ASGI server (streamable HTTP 위에 올리기 위함)
  "uvicorn[standard]>=0.32.0",
  # FastAPI (MCP 의 streamable HTTP app 마운트용)
  "fastapi>=0.120.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.3",
  "pytest-asyncio>=1.3.0",
  "testcontainers[postgres]>=4.14.2",
  "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
